+++ src/Terrabuild/Contracts/Api.fs
@@ -33,5 +33,5 @@ type IApiClient =
     abstract CompleteBuild: success:bool -> Unit
     abstract AddArtifact: project:string -> projectName:string option -> target:string -> projectHash:string -> targetHash:string  -> files:string list -> success:bool -> startedAt:DateTime -> endedAt:DateTime -> Unit
     abstract UseArtifact: projectHash:string -> hash:string -> Unit
-    abstract GetArtifact: path:string -> Uri
+    abstract GetArtifact: path:string -> operation:string -> Uri * string option
     abstract GetCommitGraph: repository:string -> commit:string -> environment:string -> CommitGraph